Initial Commit
The initial files. So far this consists on script which writes the desired SPAD-based LiDAR sensor to a .json file and another script which consists of two functions: generate_scene_conditions() and photon_numbers(). The former generates a random number of surfaces (between one and four) each at a seperate distance, their surface reflectivities, and the number of SPADs within a macropixel which that surface occupies. The latter, photon_numbers(), estimates the number of incident signal and noise photons which could be detected by each SPAD, which will later be used to distribute the photons in time.